Update italy.txt
I changed pelloux away from democratic reformer because he is the least democratic person in the whole list. He was a military strongman who tried to subvert the liberal constitution and restore power to the monarchy, seeing him as democratic reformer gave me a heart attack.

Douhet was a theorist, an early developer of air force doctrine, so he should be moved there.

Pollio was chief of staff up to the general start of the war and its first months. would have remained such and led the army instead of cadorna had he not rather suddenly died. He should be a army_chief option

things to add: the original TGW game file had lots of unused ideas portraits, including of giovanni giolitti. he was the liberal democratic prime minister for most of the years leading up to WWI and was considered a diplomatic fox: he is your smooth_talking_charmer. a more proper portrait of his should also be the democratic country leader.